requirements.txt
# データ処理
pandas>=1.5.0
numpy>=1.23.0

# データ取得
yfinance>=0.2.18
requests>=2.28.0

# 日付処理
jpholiday>=0.1.8
python-dateutil>=2.8.2

# 設定管理
pyyaml>=6.0
python-dotenv>=0.21.0

# ロギング
loguru>=0.6.0

# データベース
sqlalchemy>=2.0.0

# テスト
pytest>=7.2.0
pytest-cov>=4.0.0

# 可視化
matplotlib>=3.6.0
plotly>=5.11.0
seaborn>=0.12.0

# レポート生成
jinja2>=3.1.0
markdown>=3.4.0

# タイプヒント
typing-extensions>=4.5.0
